break up worker.go and worker_test.go into multiple files
