+++ pkg/async/redis/worker_test.go
@@ -2,11 +2,9 @@ package redis
 
 import (
 	"context"
-	"errors"
 	"testing"
 	"time"
 
-	"github.com/Azure/open-service-broker-azure/pkg/async"
 	"github.com/Azure/open-service-broker-azure/pkg/async/redis/fake"
 	"github.com/stretchr/testify/assert"
 )
@@ -417,448 +415,4 @@ func TestWorkerRunRespondsToContextCanceled(t *testing.T) {
 			"a context canceled error should have been returned, but wasn't",
 		)
 	}
-}
-
-// TestDefaultReceiveTasks tests the happy path for the defaultReceiveTasks
-// function that transplants tasks from a source queue into a destination queue.
-func TestDefaultReceiveTasks(t *testing.T) {
-	w := newWorker(redisClient).(*worker)
-
-	sourceQueueName := getDisposableQueueName()
-	destinationQueueName := getDisposableQueueName()
-
-	// Put some tasks on the source task queue
-	const taskCount int64 = 5
-	for range [taskCount]struct{}{} {
-		// Dummy tasks are fine. This test won't ever parse them.
-		err := redisClient.LPush(sourceQueueName, "foo").Err()
-		assert.Nil(t, err)
-	}
-
-	// Assert that the source queue has precisely taskCount tasks
-	sourceQueueDepth, err := redisClient.LLen(sourceQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, taskCount, sourceQueueDepth)
-
-	// Assert that the destination queue is empty
-	destinationQueueDepth, err := redisClient.LLen(destinationQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, destinationQueueDepth)
-
-	// Under nominal conditions, defaultReceiveTasks blocks until the context it
-	// is passed is canceled. Use a context that will cancel itself after 1 second
-	// to make defaultReceiveTasks STOP working so we can then examine what it
-	// accomplished.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	retCh := make(chan []byte)
-	errCh := make(chan error)
-	go w.defaultReceiveTasks(
-		ctx,
-		sourceQueueName,
-		destinationQueueName,
-		retCh,
-		errCh,
-	)
-
-	// Start another goroutine to receive and count results
-	var resCount int64
-	go func() {
-		for {
-			select {
-			case <-retCh:
-				resCount++
-			case <-ctx.Done():
-			}
-		}
-	}()
-
-	select {
-	case <-errCh:
-		assert.Fail(t, "should not have received any error, but did")
-	case <-ctx.Done():
-	}
-
-	// Assert that precisely taskCount tasks were placed onto the return channel
-	assert.Equal(t, taskCount, resCount)
-
-	// Assert that the source task queue has been drained
-	sourceQueueDepth, err = redisClient.LLen(sourceQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, sourceQueueDepth)
-
-	// Assert that the destination queue now has precisely taskCount tasks
-	destinationQueueDepth, err = redisClient.LLen(destinationQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, taskCount, destinationQueueDepth)
-}
-
-func TestDefaultExecuteTasks(t *testing.T) {
-	w := newWorker(redisClient).(*worker)
-
-	pendingTaskQueueName := getDisposableQueueName()
-	deferredTaskQueueName := getDisposableQueueName()
-	activeTaskQueueName := getActiveTaskQueueName(w.id)
-
-	// Register some jobs with the worker
-	var badJobCallCount int
-	err := w.RegisterJob(
-		"badJob",
-		func(_ context.Context, _ async.Task) ([]async.Task, error) {
-			badJobCallCount++
-			return nil, errors.New("a deliberate error")
-		},
-	)
-	assert.Nil(t, err)
-	var goodJobCallCount int
-	err = w.RegisterJob(
-		"goodJob",
-		func(_ context.Context, _ async.Task) ([]async.Task, error) {
-			goodJobCallCount++
-			return []async.Task{
-				async.NewTask("followUpJob", nil),
-				async.NewDelayedTask("followUpJob", nil, time.Minute),
-			}, nil
-		},
-	)
-	assert.Nil(t, err)
-
-	// Define some tasks
-	invalidTaskJSON := []byte("bogus")
-	unregisteredTask := async.NewTask("nonExistingJob", map[string]string{})
-	unregisteredTaskJSON, err := unregisteredTask.ToJSON()
-	assert.Nil(t, err)
-	badTask := async.NewTask("badJob", map[string]string{})
-	badTaskJSON, err := badTask.ToJSON()
-	assert.Nil(t, err)
-	goodTask := async.NewTask("goodJob", map[string]string{})
-	goodTaskJSON, err := goodTask.ToJSON()
-	assert.Nil(t, err)
-	assert.Nil(t, err)
-	tasks := [][]byte{
-		invalidTaskJSON,
-		unregisteredTaskJSON,
-		badTaskJSON,
-		goodTaskJSON,
-	}
-
-	// Put all the tasks on the worker's active task queue
-	for _, task := range tasks {
-		err := redisClient.LPush(activeTaskQueueName, task).Err()
-		assert.Nil(t, err)
-	}
-
-	// Assert that the pending task queue is empty
-	pendingTaskQueueDepth, err := redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// Assert that the deferred task queue is empty
-	deferredTaskQueueDepth, err := redisClient.LLen(deferredTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, deferredTaskQueueDepth)
-
-	// Assert that worker's active task queue has precisely len(tasks) tasks
-	activeTaskQueueDepth, err := redisClient.LLen(activeTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(len(tasks)), activeTaskQueueDepth)
-
-	// Under nominal conditions, defaultExecuteTasks blocks until the context it
-	// is passed is canceled. Use a context that will cancel itself after 1 second
-	// to make defaultExecuteTasks STOP working so we can then examine what it
-	// accomplished.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	// Put all the tasks on the defaultExecuteTasks function's input channel
-	inputCh := make(chan []byte)
-	go func() {
-		for _, task := range tasks {
-			select {
-			case inputCh <- task:
-			case <-ctx.Done():
-			}
-		}
-	}()
-
-	errCh := make(chan error)
-	go w.defaultExecuteTasks(
-		ctx,
-		inputCh,
-		pendingTaskQueueName,
-		deferredTaskQueueName,
-		errCh,
-	)
-
-	select {
-	case <-errCh:
-		assert.Fail(t, "should not have received any error, but did")
-	case <-ctx.Done():
-	}
-
-	// Assert that the pending task queue has precisely two tasks-- the
-	// unprocessable task, which should have been returned to it, and a follow-up
-	// task
-	pendingTaskQueueDepth, err = redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(2), pendingTaskQueueDepth)
-
-	// Assert that the deferred task queue has precisely one task-- a follow-up
-	// task
-	deferredTaskQueueDepth, err = redisClient.LLen(deferredTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), deferredTaskQueueDepth)
-
-	// Assert that the worker's active task queue is empty-- in all cases, the
-	// tasks should have been removed from this queue
-	activeTaskQueueDepth, err = redisClient.LLen(activeTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, activeTaskQueueDepth)
-
-	// Assert that the indicated jobs were invoked the appropriate number of times
-	assert.Equal(t, 1, badJobCallCount)
-	assert.Equal(t, 1, goodJobCallCount)
-}
-
-// TestDefaultWatchDeferredTaskWithInvalidTask tests a specific failure
-// condition for the defaultWatchDeferredTask function. The expected behavior is
-// that the task is discarded and the defaultWatchDeferredTask function adds no
-// error to its error channel.
-func TestDefaultWatchDeferredTaskWithInvalidTask(t *testing.T) {
-	w := newWorker(redisClient).(*worker)
-
-	pendingTaskQueueName := getDisposableQueueName()
-	watchedTaskQueueName := getWatchedTaskQueueName(w.id)
-
-	// Assert that the pending task queue is empty
-	pendingTaskQueueDepth, err := redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// Put an invalid task (it isn't even JSON) on the worker's watched task queue
-	invalidTaskJSON := []byte("bogus")
-	err = redisClient.LPush(watchedTaskQueueName, invalidTaskJSON).Err()
-	assert.Nil(t, err)
-
-	// Assert that queue now has precisely 1 task
-	watchedTaskQueueDepth, err := redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), watchedTaskQueueDepth)
-
-	// Under nominal conditions, defaultWatchDeferredTask could block for a very
-	// long time, unless the context it is passed is canceled. Use a context that
-	// will cancel itself after 1 second to make defaultWatchDeferredTask STOP
-	// working so we can then examine what it accomplished.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	// Call defaultWatchDeferredTask in a goroutine. Under nominal conditions,
-	// this function has the potential to run for a long time. In case the
-	// function doesn't handle the failure case we're testing properly and return
-	// quickly, we do not want to stall this test.
-	errCh := make(chan error)
-	go w.defaultWatchDeferredTask(
-		ctx,
-		invalidTaskJSON,
-		pendingTaskQueueName,
-		errCh,
-	)
-	select {
-	case <-errCh:
-		assert.Fail(t, "should not have received any error, but did")
-	case <-ctx.Done():
-	}
-
-	// Assert that the pending task queue is STILL empty
-	pendingTaskQueueDepth, err = redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// And so is the worker's watched task queue
-	watchedTaskQueueDepth, err = redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, watchedTaskQueueDepth)
-}
-
-// TestDefaultWatchDeferredTaskWithTaskWithoutExecuteTime tests a specific
-// failure condition for the defaultWatchDeferredTask function. The expected
-// behavior is that the task is discarded and the defaultWatchDeferredTask
-// function adds no error to its error channel.
-func TestDefaultWatchDeferredTaskWithTaskWithoutExecuteTime(t *testing.T) {
-	w := newWorker(redisClient).(*worker)
-
-	pendingTaskQueueName := getDisposableQueueName()
-	watchedTaskQueueName := getWatchedTaskQueueName(w.id)
-
-	// Assert that the pending task queue is empty
-	pendingTaskQueueDepth, err := redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// Put a task with no execute time on the worker's watched task queue
-	task := async.NewTask("foo", nil)
-	taskJSON, err := task.ToJSON()
-	assert.Nil(t, err)
-	err = redisClient.LPush(watchedTaskQueueName, taskJSON).Err()
-	assert.Nil(t, err)
-
-	// Assert that queue now has precisely 1 task
-	watchedTaskQueueDepth, err := redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), watchedTaskQueueDepth)
-
-	// Under nominal conditions, defaultWatchDeferredTask could block for a very
-	// long time, unless the context it is passed is canceled. Use a context that
-	// will cancel itself after 1 second to make defaultWatchDeferredTask STOP
-	// working so we can then examine what it accomplished.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	// Call defaultWatchDeferredTask in a goroutine. Under nominal conditions,
-	// this function has the potential to run for a long time. In case the
-	// function doesn't handle the failure case we're testing properly and return
-	// quickly, we do not want to stall this test.
-	errCh := make(chan error)
-	go w.defaultWatchDeferredTask(
-		ctx,
-		taskJSON,
-		pendingTaskQueueName,
-		errCh,
-	)
-	select {
-	case <-errCh:
-		assert.Fail(t, "should not have received any error, but did")
-	case <-ctx.Done():
-	}
-
-	// Assert that the pending task queue is STILL empty
-	pendingTaskQueueDepth, err = redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// And so is the worker's watched task queue
-	watchedTaskQueueDepth, err = redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, watchedTaskQueueDepth)
-}
-
-// TestDefaultWatchDeferredTaskWithLapsedTask tests that when
-// defaultWatchDeferredTask is invoked for a task whose execute time has already
-// lapsed, that task is moved IMMEDIATELY to the pending task queue.
-func TestDefaultWatchDeferredTaskWithLapsedTask(t *testing.T) {
-	w := newWorker(redisClient).(*worker)
-
-	pendingTaskQueueName := getDisposableQueueName()
-	watchedTaskQueueName := getWatchedTaskQueueName(w.id)
-
-	// Assert that the pending task queue is empty
-	pendingTaskQueueDepth, err := redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// Put a lapsed task on the worker's watched task queue
-	task := async.NewDelayedTask("foo", nil, time.Second*-1)
-	taskJSON, err := task.ToJSON()
-	assert.Nil(t, err)
-	err = redisClient.LPush(watchedTaskQueueName, taskJSON).Err()
-	assert.Nil(t, err)
-
-	// Assert that queue now has precisely 1 task
-	watchedTaskQueueDepth, err := redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), watchedTaskQueueDepth)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	// Call defaultWatchDeferredTask in a goroutine. Under nominal conditions,
-	// this function has the potential to run for a long time, although in this
-	// edge case, it should not. In case the function doesn't handle the edge case
-	// we're testing properly and return quickly, we do not want to stall this
-	// test.
-	errCh := make(chan error)
-	go w.defaultWatchDeferredTask(
-		ctx,
-		taskJSON,
-		pendingTaskQueueName,
-		errCh,
-	)
-	select {
-	case <-errCh:
-		assert.Fail(t, "should not have received any error, but did")
-	case <-time.After(time.Second):
-	}
-
-	// Assert that the pending task queue now has precisely 1 task
-	pendingTaskQueueDepth, err = redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), pendingTaskQueueDepth)
-
-	// And the worker's watched task queue is now empty
-	watchedTaskQueueDepth, err = redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, watchedTaskQueueDepth)
-}
-
-// TestDefaultWatchDeferredTaskRespondsToCanceledContext tests that when
-// defaultWatchDeferredTask is waiting for a tasks execute time to lapse, it
-// will abort if context is canceled.
-func TestDefaultWatchDeferredTaskRespondsToCanceledContext(t *testing.T) {
-	w := newWorker(redisClient).(*worker)
-
-	pendingTaskQueueName := getDisposableQueueName()
-	watchedTaskQueueName := getWatchedTaskQueueName(w.id)
-
-	// Assert that the pending task queue is empty
-	pendingTaskQueueDepth, err := redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// Put task with a future execute time on the worker's watched tasks queue
-	task := async.NewDelayedTask("foo", nil, time.Second*5)
-	taskJSON, err := task.ToJSON()
-	assert.Nil(t, err)
-	err = redisClient.LPush(watchedTaskQueueName, taskJSON).Err()
-	assert.Nil(t, err)
-
-	// Assert that queue now has precisely 1 task
-	watchedTaskQueueDepth, err := redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), watchedTaskQueueDepth)
-
-	// Use a context that will cancel itself in 1 second to put a time limit
-	// on the test. This context will be canceled BEFORE the execute time lapses.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-
-	// Call defaultWatchDeferredTask in a goroutine. Under nominal conditions,
-	// this function has the potential to run for a long time. In case the
-	// function doesn't handle context cancelation properly and return quickly, we
-	// do not want to stall this test.
-	errCh := make(chan error)
-	go w.defaultWatchDeferredTask(
-		ctx,
-		taskJSON,
-		pendingTaskQueueName,
-		errCh,
-	)
-	select {
-	case <-errCh:
-		assert.Fail(t, "should not have received any error, but did")
-	case <-ctx.Done():
-	}
-
-	// Because the context was canceled BEFORE the execute time lapsed, nothing
-	// should have changed in the queues....
-
-	// Assert that the pending task queue is STILL empty
-	pendingTaskQueueDepth, err = redisClient.LLen(pendingTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Empty(t, pendingTaskQueueDepth)
-
-	// And the worker's watched task STILL has precisely 1 task
-	watchedTaskQueueDepth, err = redisClient.LLen(watchedTaskQueueName).Result()
-	assert.Nil(t, err)
-	assert.Equal(t, int64(1), watchedTaskQueueDepth)
 }
